Created an object for the boardrecognizer process
