+++ BoardRecognizer/Makefile
@@ -1,9 +1,8 @@
 .PHONY: all clean
 
-all: recognizer
+all: protoc_middleman
 
 clean:
-	rm -f recognizer
 	rm -f protoc_middleman boardmessage_pb2.py
 	rm -f *.pyc
 
@@ -11,8 +10,3 @@ protoc_middleman: boardmessage.proto
 	protoc --python_out=. boardmessage.proto
 	@touch protoc_middleman
 
-recognizer: main.py protoc_middleman
-	@echo "Writing shortcut script host..."
-	@echo '#! /bin/sh' > recognizer
-	@echo './main.py "$$@"' >> recognizer
-	@chmod +x recognizer
